+++ pkg/adapter/prometheus/interfaces.go
@@ -1,0 +1,7 @@
+package prometheus
+
+import "net/http"
+
+type roundTripper interface {
+	http.RoundTripper
+}